+++ source/B2B.Transactions.CimMessageAdapter.Tests/Stubs/TransactionQueueDispatcherStub.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using B2B.CimMessageAdapter;
+using B2B.CimMessageAdapter.Transactions;
 
 namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests.Stubs
 {
